requirements.txt
Aadium-theme-ubuntu==0.3.4
cryptography==1.7.1
dj-database-url==0.4.2
Django==1.11.6
django-bootstrap-form==3.2.1
django-bootstrap3==9.0.0
django-cors-headers==2.1.0
django-rest-framework-roles==0.5
djangorestframework==3.6.4
drf-url-filters==0.5.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.17
keyring==10.3.1
keyrings.alt==2.2
pbr==3.1.1
py==1.4.32
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pytest==3.0.6
python-decouple==3.1
pytz==2017.2
pyxdg==0.25
SecretStorage==2.3.1
six==1.10.0
stevedore==1.25.0
unity-lens-photos==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
voluptuous==0.10.5
